requirements.txt
SpeechRecognition
pyttsx3
pyaudio
pillow
flask
flask-socketio
eventlet
openai
ollama
numpy
pyautogui
sounddevice
requests
opencv-python
face-recognition
cmake
edge-tts
pygame
PyQt5
PyQtWebEngine
PyQt5-tools
psutil
cryptography
pydub
google-generativeai
playsound
groq
websocket-client
pocketsphinx
vosk
piper-tts
yt-dlp
